pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "isds-traffic-ai"
version = "0.1.0"
description = "End-to-end traffic congestion estimation pipeline based on DeepLabV3 and YOLO11s as described in ISDS 2025 paper."
readme = "README.md"
authors = [
  {name = "Codex Agent"}
]
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
  "torch>=2.2",
  "torchvision>=0.17",
  "ultralytics>=8.2",
  "opencv-python>=4.8",
  "numpy>=1.23",
  "scipy>=1.10",
  "scikit-image>=0.22",
  "rich>=13.7",
  "pyyaml>=6.0",
  "wandb>=0.17"
]
